changes request to configurations
added changes to configurations based on comments in PR (https://github.com/Azure/azure-iot-sdk-csharp/pull/2517)
